Fix #735 refactor tools to add element in diagram
refactor :
- add exisitng element tool
- drop tool from model
Use the same rules to determine if an element could be added to a diagram.
